Split integration tests based on purpose
